Change numbers file by two distinct file with natural and integer
